MDEV-25616: Binlog event for XA COMMIT is generated without matching XA START, replication aborts
Problem:
========
If the body of an XA transaction is empty, the XA COMMIT
event will be written into the binlog; however, no other
events will be added that relate to the transaction.
The body of an XA transaction is effectively empty if
all of its statements end in error.

Fix:
====
Ensure that the COMMIT is consistent with the other XA
statements so it is also not written into the binlog